+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## 1.0a1 (unreleased)
 
+- Support `pytest` as test runner and make it default if not configured
+  otherwise.
+
 - Add `RUN_TARGET` setting to `base` domain and generate `run` target in
   `Makefile`.